Updated py files to python3

+++ CAHRIM/ActionsLib/NAOqi_apps/asr2/lib/otherlibs/pyasn1/compat/integer.py
@@ -26,7 +26,7 @@ if sys.version_info[0:2] < (3, 2) or implementation != 'CPython':
         if not octets:
             return 0
 
-        value = long(b2a_hex(ensureString(octets)), 16)
+        value = int(b2a_hex(ensureString(octets)), 16)
 
         if signed and oct2int(octets[0]) & 0x80:
             return value - (1 << len(octets) * 8)
